// src/api/types/customer/typeCstBattery.tsx
export interface typeCstBattery {
    battery_id: number;
    site_name: string;
    group_name: string;
    device_name: string;
    phonenumber: string;
    address_main: string;
    category: string;
    manufacturer: string;
    model_name: string;
    registration_date: string;
    user_name: string;
} 